pyproject.toml
[tool.poetry]
name = "discordvoicechannelnotifybot"
version = "1.0.0"
description = "Discord Bot that notifies users when they enter or leave a voice channel"
authors = ["riro110"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "==0.19.2"
beautifulsoup4 = "==4.10.0"
requests = "==2.27.1"
calmap = "==0.0.9"
numpy = "==1.22.2"
pandas = "==1.4.1"
py-cord = "==2.0.0.b1"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
isort = "^5.10.1"
mypy = "^0.931"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
